+++ order-service/src/main/java/com/shop/order/model/Order.java
@@ -1,6 +1,7 @@
 package com.shop.order.model;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 /**
@@ -20,9 +21,8 @@ public class Order {
 
     private Long userId;
 
-    // 注意：这里使用 double 类型存储金额，存在精度问题
-    // 正确的做法应该使用 BigDecimal
-    private double totalAmount;
+    // 金额使用 BigDecimal，保证精度
+    private BigDecimal totalAmount;
 
     // 订单状态：CREATED, PAYING, PAID, SHIPPED, COMPLETED, CANCELLED
     private String status;
@@ -37,7 +37,7 @@ public class Order {
     public Order() {
     }
 
-    public Order(String orderNo, Long userId, double totalAmount) {
+    public Order(String orderNo, Long userId, BigDecimal totalAmount) {
         this.orderNo = orderNo;
         this.userId = userId;
         this.totalAmount = totalAmount;
@@ -71,11 +71,11 @@ public class Order {
         this.userId = userId;
     }
 
-    public double getTotalAmount() {
+    public BigDecimal getTotalAmount() {
         return totalAmount;
     }
 
-    public void setTotalAmount(double totalAmount) {
+    public void setTotalAmount(BigDecimal totalAmount) {
         this.totalAmount = totalAmount;
     }
